@@ -1,7 +1,7 @@
 {
   "name": "apiexplorer",
-  "version": "0.5.0",
-  "description": "API Explorer is a live documentation client for Hypermedia APIs",
+  "version": "1.0.0",
+  "description": "Main API Explorer package",
   "main": "lib/index.js",
   "module": "es/index.js",
   "files": [
@@ -12,8 +12,8 @@
   ],
   "scripts": {
     "build": "nwb build-react-component",
-    "build:watch": "nodemon -x 'yarn build' -w src -e '.js'",
-    "clean": "nwb clean-module && npm clean-demo",
+    "build:watch": "nodemon -x 'npm run build' -w src -e '.js'",
+    "clean": "nwb clean-module && nwb clean-demo",
     "start": "nwb serve-react-demo",
     "test": "nwb test-react",
     "test:coverage": "nwb test-react --coverage",
@@ -21,59 +21,41 @@
   },
   "dependencies": {
     "classnames": "^2.2.5",
-    "codemirror": "^5.25.0",
-    "history": "^4.6.1",
-    "immutable": "^3.8.1",
-    "linq": "^3.0.8",
+    "codemirror": "^5.31.0",
+    "history": "^4.7.2",
+    "immutable": "^3.8.2",
+    "linq": "^3.0.9",
     "marked": "^0.3.6",
-    "moment": "^2.18.1",
-    "prop-types": "^15.5.8",
-    "react-dock": "^0.2.3",
-    "react-helmet": "^5.0.3",
-    "react-highlight": "^0.9.0",
-    "react-jsonschema-form": "^0.46.0",
-    "react-redux": "^5.0.4",
-    "react-redux-toastr": "^6.2.4",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
-    "react-router-redux": "next",
-    "redux": "^3.6.0",
+    "moment": "^2.19.1",
+    "prop-types": "^15.6.0",
+    "react-dock": "^0.2.4",
+    "react-helmet": "^5.2.0",
+    "react-highlight": "^0.10.0",
+    "react-redux": "^5.0.6",
+    "react-redux-toastr": "^7.1.6",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
+    "react-router-redux": "^5.0.0-alpha.8",
+    "redux": "^3.7.2",
     "redux-thunk": "^2.2.0",
-    "semantic-ui-react": "^0.68.0",
-    "styled-components": "^1.4.5",
+    "semantic-ui-react": "^0.76.0",
+    "styled-components": "^2.2.3",
     "swagger-converter": "^1.4.1",
-    "urijs": "^1.18.10"
+    "urijs": "^1.19.0"
   },
   "peerDependencies": {
     "react": "15.x"
   },
   "devDependencies": {
-    "nwb": "0.15.x",
-    "react": "^15.5.3",
-    "react-dom": "^15.5.3"
+    "nwb": "0.19.x",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   },
-  "author": "Sky UK Limited",
-  "homepage": "https://github.com/sky-uk/api-explorer",
-  "license": "SEE LICENSE IN LICENSE",
-  "repository": "https://github.com/sky-uk/api-explorer",
+  "author": "",
+  "homepage": "",
+  "license": "MIT",
+  "repository": "",
   "keywords": [
-    "api",
-    "explorer",
-    "hypermedia",
-    "documentation",
-    "tool"
-  ],
-  "bugs": {
-    "url": "https://github.com/sky-uk/api-explorer/issues"
-  },
-  "maintainers": [
-    {
-      "name": "Carlos Guedes",
-      "email": "carlos.guedes@sky.uk"
-    },
-    {
-      "name": "Daniel Correia",
-      "email": "daniel.correia@sky.uk"
-    }
+    "react-component"
   ]
 }
