@@ -25,14 +25,14 @@ class HowToConfigureAPIExplorer extends Component {
             <p>It seems that you have an <strong>empty</strong> API Explorer configuration. Check out a sample configuration.</p>
             <br />
             <div style={{textAlign: 'left', marginLeft: '20%', marginRight: '20%'}}>
-              <Highlight className='JavaScript'>{`APIExplorer
+              {/* <Highlight className='JavaScript'>{`APIExplorer
   .addAPI('petstore', 'swagger2', 'https://api.swaggerhub.com/apis/anil614sagar/petStore/1.0.0', c => {
     c.addHeader('X-Foo', 'Some Value')
     c.addHeader('X-Bar', 'Another Value')
   })
   .configCORS({ credentials: 'omit' })
   .start()`}
-              </Highlight>
+              </Highlight> */}
             </div>
             <a href='https://github.com/sky-uk/api-explorer'>https://github.com/sky-uk/api-explorer</a>
 
