@@ -76,12 +76,12 @@ class TryOutWidgetTabExecuter extends Component {
   }
 }
 
-TryOutWidgetTabExecuter.propTypes = {
-  requestFormat: PropTypes.string,
-  requestFormats: PropTypes.array,
-  requestInProgress: PropTypes.bool.isRequired,
-  onExecuteRequest: PropTypes.func.isRequired,
-  onValidateParameters: PropTypes.func.isRequired
-}
+// TryOutWidgetTabExecuter.propTypes = {
+//   requestFormat: PropTypes.string,
+//   requestFormats: PropTypes.array,
+//   requestInProgress: PropTypes.bool.isRequired,
+//   onExecuteRequest: PropTypes.func.isRequired,
+//   onValidateParameters: PropTypes.func.isRequired
+// }
 
 export default TryOutWidgetTabExecuter
